Move pdns_i*() from misc.hh to dnsname.hh. NFCI
This will avoid future chicken-and-egg problems.
